+++ tools/scripts/generate-rule-docs.ts
@@ -1,8 +1,8 @@
 import type { TSESLint } from '@typescript-eslint/utils';
-import { readdirSync, readFileSync, writeFileSync } from 'fs';
 import { compile } from 'json-schema-to-typescript';
 import traverse from 'json-schema-traverse';
-import { join, relative } from 'path';
+import { readdirSync, readFileSync, writeFileSync } from 'node:fs';
+import { join, relative } from 'node:path';
 import { format, resolveConfig } from 'prettier';
 import ts from 'typescript';
 import { SPECIAL_UNDERLINE_CHARS } from '../../packages/utils/src/convert-annotated-source-to-failure-case';
@@ -38,7 +38,9 @@ const testDirs = readdirSync(testDirsDir);
       testCasesFilePath,
     } = ruleData;
 
-    const docs = ruleData.ruleConfig.meta.docs as TSESLint.RuleMetaDataDocs;
+    const docs = ruleData.ruleConfig.meta.docs as TSESLint.RuleMetaDataDocs<
+      unknown[]
+    >;
     const { description } = docs;
 
     let schemaAsInterface = '';
@@ -109,14 +111,18 @@ const testDirs = readdirSync(testDirsDir);
 
 ${
   deprecated
-    ? `## ⚠️ THIS RULE IS DEPRECATED\n\nPlease use ${(replacedBy || [])
-        .map(
-          (r: string) =>
-            `\`@angular-eslint/${
-              plugin === 'eslint-plugin-template' ? 'template/' : ''
-            }${r}\``,
-        )
-        .join(', ')} instead.\n\n---\n\n`
+    ? `## ⚠️ THIS RULE IS DEPRECATED\n\n${
+        replacedBy
+          ? `Please use ${(replacedBy || [])
+              .map(
+                (r: string) =>
+                  `\`@angular-eslint/${
+                    plugin === 'eslint-plugin-template' ? 'template/' : ''
+                  }${r}\``,
+              )
+              .join(', ')} instead.`
+          : ''
+      }\n\n---\n\n`
     : ''
 }${description}